requirements.txt
robin_stocks
dotenv
discord.py
schedule
public-invest-api